@@ -5,24 +5,24 @@
     "license": "MIT",
     "type": "library",
     "require": {
-        "php": "^7.4|^8.0",
-        "anahkiasen/former": "^4.5",
+        "php": "^8.0",
+        "anahkiasen/former": "^4.6",
         "atrox/haikunator": "^1.3",
         "fideloper/proxy": "^4.4",
-        "guzzlehttp/guzzle": "^6.5|^7",
+        "guzzlehttp/guzzle": "^7.3",
         "haydenpierce/class-finder": "^0.4",
-        "laravel/framework": "^8",
-        "laravel/helpers": "^1.3",
-        "laravel/passport": "^10.0",
-        "laravel/socialite": "^5.0",
-        "mll-lab/graphql-php-scalars": "^3.1",
-        "nuwave/lighthouse": "4.18.*",
-        "paragonie/certainty": "^2.7",
-        "ramsey/uuid": "^3.9|^4.1",
-        "sentry/sentry-laravel": "^2.4"
+        "laravel/framework": "^8.42",
+        "laravel/helpers": "^1.4",
+        "laravel/passport": "^10.1",
+        "laravel/socialite": "^5.2",
+        "mll-lab/graphql-php-scalars": "^4.1",
+        "nuwave/lighthouse": "5.10.*",
+        "paragonie/certainty": "^2.8",
+        "ramsey/uuid": "^4.1",
+        "sentry/sentry-laravel": "^2.5"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "2.18.*"
+        "friendsofphp/php-cs-fixer": "2.19.*"
     },
     "autoload": {
         "psr-4": {
