@@ -3,7 +3,10 @@
 use IronGate\Chief\Http\Controllers;
 
 Route::group(config('chief.routes.api'), function () {
-    Route::match(['get', 'post'], 'graphql', [Controllers\API\GraphQL::class, 'query'])->name('api');
+    Route::match(['get', 'post'], 'graphql', Controllers\API\GraphQL::class)->middleware([
+        Nuwave\Lighthouse\Support\Http\Middleware\AcceptJson::class,
+        IronGate\Chief\GraphQL\Middleware\AutoAuthenticate::class,
+    ])->name('api');
 
     Route::get('graphql/schema', [Controllers\API\GraphQL::class, 'schema'])->name('api.schema');
 
