@@ -3,7 +3,10 @@
 use IronGate\Chief\Http\Controllers;
 
 Route::group(config('chief.routes.web-api'), function () {
-    Route::match(['get', 'post'], 'graphql/web', [Controllers\API\GraphQL::class, 'queryWeb'])->name('api.web');
+    Route::match(['get', 'post'], 'graphql/web', Controllers\API\GraphQL::class)->middleware([
+        Nuwave\Lighthouse\Support\Http\Middleware\AcceptJson::class,
+        IronGate\Chief\GraphQL\Middleware\AutoAuthenticate::class,
+    ])->name('api.web');
 
     Route::view('playground', 'chief::api.playground.' . config('chief.graphql.playground'))->name('api.playground');
 });
